UPDATE README
We believe that the loss of the "What is NPOI" section is not a work error. Due to the authorization of the upstream project author, removing this information does not violate the agreement or license. Because the upstream project has fully explained the relationship between NPOI and Apache POI. Here we will add it back.